% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{intr_csv_writer}
\alias{intr_csv_writer}
\title{Write or append to csv}
\usage{
intr_csv_writer(path_, df)
}
\arguments{
\item{path_}{character. Path to desired csv.}

\item{df}{data.frame. Data to write to disk.}
}
\description{
Write or append to csv
}
