% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{intr_queue}
\alias{intr_queue}
\title{Queue file manager.}
\usage{
intr_queue(path_, add_ = NULL, rm_ = FALSE, read_ = FALSE)
}
\arguments{
\item{path_}{character. A single file path to the txt file. Make sure this is consistent each time the file loads.}

\item{add_}{character. One or more paths in a character vector. Set to NULL when nothing to add.}

\item{rm_}{logical. Determines whether to delete the top line or not.}

\item{read_}{logical. Read the first line of the file?}
}
\value{
character.
}
\description{
Manages a txt file that contains a list of file paths. It can do three things in any combination: 1) Read the first line and return it (which it does by default), 2) Append paths to the bottom of the txt, and 3) remove the first line of the text file. This function does not read the whole file. It will create an empty txt document if none exists.
}
