% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{k_inventory}
\alias{k_inventory}
\title{Single directory inventory maker}
\usage{
k_inventory(
  dir_,
  ignore_ = NULL,
  pattern_ = NULL,
  min_size = 0,
  include_hid = FALSE
)
}
\arguments{
\item{dir_}{character. Normalized path to primary directory.}

\item{ignore_}{character. Paths to directories to ignore.}

\item{pattern_}{character. Usually file extensions to focus upon, see \link[base]{list.files}.}

\item{min_size}{numeric. The minimum file size (in bytes) to inventory.}

\item{include_hid}{logical. Whether or not to include directories and files that are hidden, ie. that begin with a period.}
}
\value{
list
}
\description{
Takes a path to a directory and inventories the contents. This function looks for directories and files. It returns a list with two elements, a character vector of directories, and a data.frame containing some basic information about files including the path, size, creation time, etc. Importantly, this function is non-recursive.
}
