% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{k_user_ls}
\alias{k_user_ls}
\title{List, retrieve metadata, and store information about all user directories and files}
\usage{
k_user_ls(
  user_ = NULL,
  ignore_library = TRUE,
  ignore_apps = TRUE,
  ignore_hidden = TRUE,
  min_size = 1000,
  include_icloud = TRUE,
  savepath = "~/Downloads",
  pattern_ = NULL,
  date_ = NULL
)
}
\arguments{
\item{user_}{character. Set origin user's name.}

\item{ignore_library}{logical. Ignore user /Library.}

\item{ignore_apps}{logical. Ignore user /Applications.}

\item{ignore_hidden}{logical. Ignore hidden files (which begin with period).}

\item{min_size}{numeric. Ignore files under a certain size (bytes).}

\item{include_icloud}{logical. Include iCloud files. This will probably significantly increase scanning time.}

\item{savepath}{character. The file path to the output directory. By default, it saves to ~/Downloads.}

\item{pattern_}{character. Useful for narrowing down certain filetypes, eg. ".pdf$". See pattern in arguments for \link[base]{list.files}.}

\item{date_}{character. The date that should be appended to any files. By default it does \code{format(Sys.time(), "\%Y_\%m_\%d")}.}
}
\value{
tibble
}
\description{
List, retrieve metadata, and store information about all user directories and files
}
