% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kondo-package.R
\docType{package}
\name{kondo-package}
\alias{kondo}
\alias{kondo-package}
\title{kondo: Organize file system on macOS.}
\description{
Organize file system on macOS.
}
\author{
\strong{Maintainer}: Kenji Tomari \email{ktomari@ucdavis.edu}

}
\keyword{internal}
