% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{k_dupes}
\alias{k_dupes}
\title{Identify duplicates files in tibble contain columns \code{path} and \code{size}.}
\usage{
k_dupes(tb, rm.unique = TRUE)
}
\arguments{
\item{tb}{data.frame/tibble. The data.frame returned by \code{k_ls}.}

\item{rm.unique}{logical. Whether or not to remove unique values. By default, it is \code{TRUE}.}
}
\value{
tibble
}
\description{
If hash is present in the \code{md5}column, then it is duplicate!
}
