% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{k_dupes}
\alias{k_dupes}
\title{Identify duplicates files in tibble contain columns \code{path} and \code{size}.}
\usage{
k_dupes(tb)
}
\arguments{
\item{tb}{}
}
\value{
tibble
}
\description{
If hash is present in the \code{md5}column, then it is duplicate!
}
