% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{k_ls}
\alias{k_ls}
\title{Create tibble of file list with information.}
\usage{
k_ls(dir_)
}
\arguments{
\item{dir_}{string. Path to directory.}
}
\value{
tibble
}
\description{
Create tibble of file list with information.
}
\examples{
k_ls(dir_ = "~/Downloads")
}
